% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancorpus.R
\name{cleancorpus}
\alias{cleancorpus}
\title{Cleancorpus function}
\usage{
cleancorpus(corpus)
}
\arguments{
\item{corpus}{A set of unstructured texts in a dataframe}
}
\value{
A cleaned corpus dataframe
}
\description{
This function is a part of the dendrogrammer function, but can be used on its own.
This function will take a corpus and clean it to be used for other forms of text mining
by removing capitalization, punctuation, whitespace, numbers, and stopwords
}
\examples{
load("~/Dendro/data/airbnbdf.rda")
corpus <- DataframeSource(airbnbdf)
corpus <- VCorpus(corpus)
corpus <- cleancorpus(corpus)
}
