% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro-package.R
\docType{package}
\name{dendro-package}
\alias{dendro}
\alias{dendro-package}
\title{dendro: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/dendro}
  \item \url{https://mikemahoney218.github.io/dendro/}
  \item Report bugs at \url{https://github.com/mikemahoney218/dendro/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
