% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dendrogrammer.R
\name{Dendrogrammer}
\alias{Dendrogrammer}
\title{Dendrogrammer function}
\usage{
Dendrogrammer(dataframe, x)
}
\arguments{
\item{dataframe}{A dataframe containing the columns "doc_id" and "text"}

\item{x}{A numeric value between 0 and 1 to write in your own sparsity threshold}
}
\value{
A dendrogram plot
}
\description{
Dendrogrammer function
}
\examples{
load("~/Dendro/data/airbnbdf.rda")
Dendrogrammer(airbnbdf, 0.97)

}
