% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_depth.R
\name{distance_to_target}
\alias{distance_to_target}
\title{Calculate distance between two nodes in a tree.}
\usage{
distance_to_target(x, ...)
}
\arguments{
\item{x}{The tree to calculate distances within.}

\item{from}{The index to calculate distance "from".}

\item{to}{The index to calculate distance "to".}
}
\value{
Distance between nodes as a numeric scalar. If the nodes are not
nested, returns NA. If \code{to} is above \code{from} in the tree, distance will be
negative.
}
\description{
Calculate distance between two nodes in a tree.
}
\examples{
tree() |>
  add_node() |>
  add_node(parent = 1L) |>
  add_node(parent = 2L) |>
  distance_to_target(1, 3)

}
