% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_node.R
\name{drop_node}
\alias{drop_node}
\title{Drop a node (and all its descendants) from a tree}
\usage{
drop_node(x, ...)
}
\arguments{
\item{x}{The tree to drop nodes from.}

\item{node_index}{The node to drop.}
}
\value{
x, without the specified node and descendants.
}
\description{
Drop a node (and all its descendants) from a tree
}
\examples{
tree() |>
  add_node() |>
  add_node(parent = 1L) |>
  drop_node(2)

tree() |>
  add_node() |>
  add_node(parent = 1L) |>
  drop_node(1)

}
