% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fh.tab}
\alias{fh.tab}
\title{Transform .fh file to special .txt format}
\usage{
fh.tab(fails,start="bark")
}
\arguments{
\item{fails}{a name of .fh file to tranform}

\item{start}{dircetion of measuraments - from "bark" or from "pith"}
}
\value{
A .txt file in working directory
}
\description{
This function takes .fh file from working directory, transforms it to special format .txt file and
 exports it to working directory
}
\examples{
#fh.tab("file.name.fh")
}

