% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_children.R
\name{get_children}
\alias{get_children}
\title{Get the children slot from a single node in a tree}
\usage{
get_children(x, ...)
}
\arguments{
\item{x}{The tree to extract the node from.}

\item{node_index}{The node index to extract children from.}
}
\value{
The children of node \code{node_index}, as an integer vector.
}
\description{
Get the children slot from a single node in a tree
}
\examples{
tree() |>
  add_node() |>
  add_node(parent = 1L) |>
  add_node(parent = 2L) |>
  get_children(1L)

}
