% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_node_values}
\alias{get_node_values}
\title{Extract nodes values from a tree}
\usage{
get_node_values(x, ...)
}
\arguments{
\item{x}{The tree to extract node values from.}

\item{node_indices}{The index of nodes to extract from the tree.}
}
\value{
A list of node values.
}
\description{
Extract nodes values from a tree
}
\examples{
tree() |>
  add_node() |>
  add_node(parent = 1L, value = list("value one!")) |>
  add_node(parent = 2L, value = list("value two!")) |>
  get_node_values(c(2, 3))

}
