% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{node}
\alias{node}
\title{Node objects for collecting within trees}
\usage{
node(
  index = class_missing,
  value = class_missing,
  parent = class_missing,
  children = class_missing
)
}
\arguments{
\item{index}{The index of the node in the list of nodes stored in the tree
object. Calculated automatically when using \code{\link[=add_node]{add_node()}}.}

\item{value}{Any data or other attributes associated with the node.}

\item{parent}{Integer: the index of the node's parent. Only one root node is
allowed per tree.}

\item{children}{Integer: the index of any children of the node. Leaf nodes
should have a value of \code{NA_integer_}.}
}
\value{
An object of class \verb{"dendro::node", "S7_object"}
}
\description{
These objects aren't particularly useful on their own, and are usually
called via \code{\link[=add_node]{add_node()}} to automatically assign a unique index.
}
\examples{
node(0L)

node(0L, list(a = 2))

}
