% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_depth.R
\name{node_depth}
\alias{node_depth}
\title{Calculate depth of a node in the tree}
\usage{
node_depth(x, ...)
}
\arguments{
\item{x}{The tree to calculate node depth in.}

\item{node_index}{The index of the node to calculate depth for}
}
\value{
The depth of the node as a numeric scalar.
}
\description{
Calculate depth of a node in the tree
}
\examples{
tree() |>
  add_node() |>
  add_node(parent = 1L) |>
  add_node(parent = 2L) |>
  node_depth(3)

}
