\name{ozols}
\alias{ozols}
\docType{data}
\title{
ozols - raw tree-ring widths values
}
\description{
This dataset contains 18 tree-ring width series of Quercus robur from Jelgava sampling site in Latvia and coresponding year in first column.
}
\usage{data(ozols)}
\format{
  A data frame with 110 observations on the following 19 variables.
  \describe{
    \item{\code{gads}}{a numeric vector}
    \item{\code{JEL1}}{a numeric vector}
    \item{\code{JEL2}}{a numeric vector}
    \item{\code{JEL3}}{a numeric vector}
    \item{\code{JEL4}}{a numeric vector}
    \item{\code{JEL5}}{a numeric vector}
    \item{\code{JEL6}}{a numeric vector}
    \item{\code{JEL7}}{a numeric vector}
    \item{\code{JEL8}}{a numeric vector}
    \item{\code{JEL9}}{a numeric vector}
    \item{\code{JEL10}}{a numeric vector}
    \item{\code{JEL11}}{a numeric vector}
    \item{\code{JEL12}}{a numeric vector}
    \item{\code{JEL13}}{a numeric vector}
    \item{\code{JEL14}}{a numeric vector}
    \item{\code{JEL21}}{a numeric vector}
    \item{\code{JEL22}}{a numeric vector}
    \item{\code{JEL23}}{a numeric vector}
    \item{\code{JEL24}}{a numeric vector}
  }
}

\examples{
data(ozols)
str(ozols)
}
\keyword{datasets}
