% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointer.year.R
\name{pointer.year}
\alias{pointer.year}
\title{Calculate pointer-year values from raw tree-ring widths}
\usage{
pointer.year(x)
}
\description{
This function takes object of raw tree-ring widths measuraments and calculates stand pointer-year 
 values for each year of original data.
 
 @param x a data frame with year values in the first column and raw tree ring measuraments in all other columns
 @usage pointer.year(x)
 @return A data frame with year values in the first column and the calculated stand pointer-year values in the second column
 @examples
 data(ozols)
 point<-pointer.year(ozols)
 point[1:10,]
}

