% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.supr.R
\name{rel.supr}
\alias{rel.supr}
\title{Calculate release-supression values from raw tree-ring widths}
\usage{
rel.supr(gadsk, sliek = 0.5)
}
\description{
This function takes object of raw tree-ring widths measuraments and calculates stand release-suppression values for 10 year interval 
 values for each year of original data.
 
 @param gadsk a data frame with year values in the first column and raw tree ring measuraments in all other columns
 @param sliek a value from 0 to 1 that shows how big increase/decrease should be to count it as release or suppression
 @usage rel.supr(gadsk,sliek=0.5)
 @return A data frame with year values in the first column, number and percentage of trees with positive and negative change and total number of trees
 @examples
 data(ozols)
 rel.values<-rel.supr(ozols,sliek=0.3)
 rel.values[1:10,]
}

