% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trans.wd}
\alias{trans.wd}
\title{Transform WinDendro measurament file of version 3}
\usage{
trans.wd(file,r.par="RINGWIDTH")
}
\arguments{
\item{file}{a name of WinDendro file}

\item{r.par}{what kind of measuraments to extract - "RINGWIDTH","EARLYWIDTH" or "LATEWIDTH"}
}
\value{
A data frame with year in first column and measuraments in all other columns
}
\description{
This function takes measurament file of WinDendro programm (version 3 file) and extracts measuraments,
 years and names, and makes data frame where each tree-ring width series are in seperate column.
}
\examples{
#trans.wd("file.name.txt",r.par="RINGWIDTH")
}

