% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tree}
\alias{tree}
\title{Tree objects}
\usage{
tree(nodes = class_missing)
}
\arguments{
\item{nodes}{A list of nodes, as created via \code{\link[=node]{node()}}, which form a tree.}
}
\value{
An object of class \verb{"dendro::tree", "S7_object"}
}
\description{
This is the primary class implemented in this package.
}
\examples{
tree()
tree(list(node(1L)))

}
