% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh.tab.R
\name{fh.tab}
\alias{fh.tab}
\title{Transform .fh file to special .txt format}
\usage{
fh.tab(fails, start = "bark")
}
\description{
This function takes .fh file from working directory, transforms it to special format .txt file and
 exports it to working directory
 
 @param fails a name of .fh file to tranform
 @param start dircetion of measuraments - from "bark" or from "pith"
 @usage fh.tab(fails,start="bark")
 @return A .txt file in working directory
 @examples
 #fh.tab("file.name.fh")
}

