% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pointer.year}
\alias{pointer.year}
\title{Calculate pointer-year values from raw tree-ring widths}
\usage{
pointer.year(x)
}
\arguments{
\item{x}{a data frame with year values in the first column and raw tree ring measuraments in all other columns}
}
\value{
A data frame with year values in the first column and the calculated stand pointer-year values in the second column
}
\description{
This function takes object of raw tree-ring widths measuraments and calculates stand pointer-year
 values for each year of original data.
}
\examples{
data(ozols)
 point<-pointer.year(ozols)
 point[1:10,]
}

