% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rel.supr}
\alias{rel.supr}
\title{Calculate release-supression values from raw tree-ring widths}
\usage{
rel.supr(gadsk,sliek=0.5)
}
\arguments{
\item{gadsk}{a data frame with year values in the first column and raw tree ring measuraments in all other columns}

\item{sliek}{a value from 0 to 1 that shows how big increase/decrease should be to count it as release or suppression}
}
\value{
A data frame with year values in the first column, number and percentage of trees with positive and negative change and total number of trees
}
\description{
This function takes object of raw tree-ring widths measuraments and calculates stand release-suppression values for 10 year interval
 values for each year of original data.
}
\examples{
data(ozols)
 rel.values<-rel.supr(ozols,sliek=0.3)
 rel.values[1:10,]
}

