% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.wd.R
\name{trans.wd}
\alias{trans.wd}
\title{Transform WinDendro measurament file of version 3}
\usage{
trans.wd(file, r.par = "RINGWIDTH")
}
\description{
This function takes measurament file of WinDendro programm (version 3 files only at this point) and extracts measuraments,
 years and names, and makes data frame where each tree-ring width series are in seperate column.
 
 @param file a name of WinDendro file
 @param r.par what kind of measuraments to extract - "RINGWIDTH","EARLYWIDTH" or "LATEWIDTH"
 @usage trans.wd(file,r.par="RINGWIDTH")
 @return A data frame with year in first column and measuraments in all other columns
 @examples
 #trans.wd("file.name.txt",r.par="RINGWIDTH")
}

